\name{ConversionTableForMunicipalities}
\alias{ConversionTableForMunicipalities}
\title{ConversionTableForMunicipalities
Used to create conversion table between municipality names and codes}
\usage{
  ConversionTableForMunicipalities()
}
\value{
  conversion table
}
\description{
  ConversionTableForMunicipalities Used to create
  conversion table between municipality names and codes
}
\examples{
#
}
\author{
  Leo Lahti \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

