\name{ConvertMunicipalityCodes}
\alias{ConvertMunicipalityCodes}
\title{Convert between municipality codes and names}
\usage{
  ConvertMunicipalityCodes(ids = NULL, names = NULL)
}
\arguments{
  \item{ids}{NULL}

  \item{names}{NULL}
}
\value{
  Depending on the input. Converted id or name vector, or
  full conversion table.
}
\description{
  Convert between municipality codes and names
}
\examples{
# conversion.table <- ConvertMunicipalityCodes()
}
\author{
  Leo Lahti \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

