\name{ConvertMunicipalityNames}
\alias{ConvertMunicipalityNames}
\title{Convert municipality names into standard versions}
\usage{
  ConvertMunicipalityNames(municipality.names)
}
\arguments{
  \item{municipality.names}{municipality names to convert}
}
\value{
  standardized municipality names
}
\description{
  Convert municipality names into standard versions
}
\examples{
# tmp <- ConvertMunicipalityNames(municipality.names)
}
\author{
  Leo Lahti \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

