\name{GetEdustajanAanestykset}
\alias{GetEdustajanAanestykset}
\title{Hakee tietyn kansanedustajan aanestykset}
\usage{
  GetEdustajanAanestykset(edustaja)
}
\arguments{
  \item{edustaja}{edustajan nimi muodossa Sukunimi Etunimi}
}
\value{
  data.frame
}
\description{
  Hakee tietyn kansanedustajan aanestykset
}
\examples{
# paavo <- GetEdustajanAanestykset('Lipponen Paavo')
}
\author{
  Juuso Haapanen \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{eduskunta}

