\name{GetMunicipalElectionData2004}
\alias{GetMunicipalElectionData2004}
\title{GetMunicipalElectionData2004}
\usage{
  GetMunicipalElectionData2004(which = "election.statistics")
}
\arguments{
  \item{which}{Indicate which of the available Statistics
  Finland data sets to parse.}
}
\value{
  data.frame
}
\description{
  Get municipal election data from Statistics Finland 2012
}
\details{
  Taulukot tilastossa: 5. Kunnallisvaalit 2004 -
  vaalitulos, aanestaminen
  http://pxweb2.stat.fi/database/StatFin/vaa/kvaa/2004_05/2004_05_fi.asp
}
\examples{
#
}
\author{
  Leo Lahti \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

