\name{GetMunicipalElectionData2008}
\alias{GetMunicipalElectionData2008}
\title{GetMunicipalElectionData2008}
\usage{
  GetMunicipalElectionData2008(which = "election.statistics")
}
\arguments{
  \item{which}{Indicate which of the available Statistics
  Finland data sets to parse. Options:}
}
\value{
  data.frame
}
\description{
  Get municipal election data
}
\examples{
#
}
\author{
  Leo Lahti \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

