\name{GetMunicipalityInfo}
\alias{GetMunicipalityInfo}
\alias{get.municipality.info}
\title{Get information of Finnish municipalities from Statistics Finland 2012
and Maanmittauslaitos (C) MML 2011. For details of MML data, see
help(GetShapeMML).}
\usage{
  GetMunicipalityInfo(...)
}
\arguments{
  \item{...}{Arguments to be passed}
}
\value{
  A data frame with municipality data
}
\description{
  Get information of Finnish municipalities from Statistics
  Finland 2012 and Maanmittauslaitos (C) MML 2011. For
  details of MML data, see help(GetShapeMML).
}
\examples{
# municipality.info <- GetMunicipalityInfo()
}
\author{
  Leo Lahti \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

