\name{GetMunicipalityInfoMML}
\alias{GetMunicipalityInfoMML}
\title{Get information of Finnish municipalities from Land Survey Finland 2012.
(C) Maanmittauslaitos MML 2012. For details of MML data, see
help(GetShapeMML).}
\usage{
  GetMunicipalityInfoMML()
}
\value{
  A data frame with municipality data
}
\description{
  Get information of Finnish municipalities from Land
  Survey Finland 2012. (C) Maanmittauslaitos MML 2012. For
  details of MML data, see help(GetShapeMML).
}
\examples{
# tab <- GetMunicipalityInfoMML()
}
\author{
  Leo Lahti \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

