\name{GetMunicipalityInfoStatFi}
\alias{GetMunicipalityInfoStatFi}
\title{Get information of Finnish municipalities from Statistics Finland 2013}
\usage{
  GetMunicipalityInfoStatFi(verbose = TRUE, ...)
}
\arguments{
  \item{verbose}{verbose}

  \item{...}{Arguments to be passed}
}
\value{
  A data frame with municipality data
}
\description{
  Get information of Finnish municipalities from Statistics
  Finland 2013
}
\examples{
# tmp <- GetMunicipalityInfo()
}
\author{
  Leo Lahti \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

