\name{GetParliamentaryElectionData}
\alias{GetParliamentaryElectionData}
\title{GetParliamentaryElectionData}
\usage{
  GetParliamentaryElectionData(level)
}
\arguments{
  \item{level}{Indicate whether to get data at the level of
  municipality or election.region}
}
\value{
  data.frame
}
\description{
  Get parliamentary election data at selected regional
  level.
}
\examples{
#
}
\author{
  Leo Lahti \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

