\name{GetVaalipiiri}
\alias{GetVaalipiiri}
\title{Load Vaalipiiri information
Useful for mapping election data to other municipality information}
\usage{
  GetVaalipiiri(...)
}
\arguments{
  \item{...}{Arguments to be passed}
}
\value{
  data.frame listing election regions (Vaalipiiri), region
  IDs (Aluenumero) and municipalities (Alue)
}
\description{
  Load Vaalipiiri information Useful for mapping election
  data to other municipality information
}
\author{
  Juuso Parkkinen and Leo Lahti
  \email{louhos@googlegroups.com}
}

