\name{haeHakusanalla}
\alias{haeHakusanalla}
\title{Hakee hakusanalla aanestyksia eduskuntarajapinnasta}
\usage{
  haeHakusanalla(hakusana)
}
\arguments{
  \item{hakusana}{string}
}
\value{
  data.frame
}
\description{
  Hakee hakusanalla aanestyksia eduskuntarajapinnasta
}
\examples{
# luonto <- haeHakusanalla('luonto')
}
\author{
  Juuso Haapanen \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{eduskunta}

