\name{LouhosStoragePath}
\alias{LouhosStoragePath}
\title{LouhosStoragePath}
\usage{
  LouhosStoragePath()
}
\value{
  URL for Louhos data
}
\description{
  Arguments: ... Arguments to pass
}
\details{
  Return:
}
\examples{
# url <- LouhosStoragePath()
}
\author{
  Leo Lahti \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

