\name{ReadAllCandidates}
\alias{ReadAllCandidates}
\title{Description:
Wrapper function for ReadCandidates that gets all 14 districts and returns
all data in a single data frame.}
\usage{
  ReadAllCandidates(cache = NA)
}
\arguments{
  \item{cache}{character directory path to location where
  files are cached}
}
\value{
  Data frame
}
\description{
  Description: Wrapper function for ReadCandidates that
  gets all 14 districts and returns all data in a single
  data frame.
}
\examples{
#
}
\author{
  Joona Lehtomaki \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

