\name{ReadAllParties}
\alias{ReadAllParties}
\title{Description:
Wrapper function for ReadParties that gets all 14 districts and returns
all data in a single data frame.}
\usage{
  ReadAllParties(cache = NA)
}
\arguments{
  \item{cache}{character directory path to location where
  files are cached}
}
\value{
  Data frame
}
\description{
  Description: Wrapper function for ReadParties that gets
  all 14 districts and returns all data in a single data
  frame.
}
\examples{
#
}
\author{
  Leo Lahti \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

