\docType{package}
\name{sorvi-package}
\alias{sorvi}
\alias{sorvi-package}
\title{Algorithmic Tools for Open Data in Finland}
\description{
  Brief summary of the sorvi package
}
\details{
  \tabular{ll}{ Package: \tab sorvi\cr Type: \tab
  Package\cr Version: \tab See sessionInfo() or DESCRIPTION
  file\cr Date: \tab 2010-2013\cr License: \tab FreeBSD\cr
  LazyLoad: \tab yes\cr }

  Algorithmic tools for Open Data in Finland
}
\examples{
library(sorvi)
}
\author{
  Leo Lahti, Juuso Parkkinen and Joona Lehtomaki
  \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
  \url{http://louhos.github.com/sorvi}
}
\keyword{package}

